requirements.txt
certifi==2022.6.15
chardet==3.0.4
click==8.1.3
Flask==1.1.1
Flask-Classful==0.14.2
idna==2.8
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
json5==0.8.4
jsonpickle==1.4.1
MarkupSafe==2.1.1
p2pnetwork==0.0.3
# pycrypto==2.6.1
pycryptodome==3.15.0
requests==2.22.0
urllib3==1.25.11
Werkzeug==2.2.1
zipp==3.8.1
